+++ frontend/controllers/MensajesPrivadosController.php
@@ -16,6 +16,7 @@ use yii\web\NotFoundHttpException;
 class MensajesPrivadosController extends Controller
 {
     use \common\utilities\Permisos;
+    use \common\traitrollers\CommonIndex;
     /**
      * @inheritdoc
      */
@@ -39,7 +40,8 @@ class MensajesPrivadosController extends Controller
      */
     public function actionIndex()
     {
-        return $this->inbox(['receptor_id' => Yii::$app->user->id, 'del_r' => false], 'index');
+        $model = new MensajesPrivadosSearch();
+        return $this->commonIndex($model, ['receptor_id' => Yii::$app->user->id, 'del_r' => false], 'index');
     }
 
     /**
@@ -48,25 +50,8 @@ class MensajesPrivadosController extends Controller
      */
     public function actionSent()
     {
-        return $this->inbox(['emisor_id' => Yii::$app->user->id, 'del_e' => false], 'sent');
-    }
-
-    /**
-     * En común con la acción index y sent.
-     * @param  array  $where where de la query
-     * @param  string $name  nombre de la acción
-     * @return mixed
-     */
-    private function inbox($where, $name)
-    {
-        $searchModel = new MensajesPrivadosSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $dataProvider->query->where($where);
-
-        return $this->render($name, [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+        $model = new MensajesPrivadosSearch();
+        return $this->commonIndex($model, ['emisor_id' => Yii::$app->user->id, 'del_e' => false], 'sent');
     }
 
     /**
